Use typed SignalR hub for movie updates
