--- Movie.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-public class Class1
-{
-	public Class1()
-	{
-	}
-}
