--- MovieHub.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.AspNetCore.SignalR;
-
-public class MovieHub : Hub
-{
-    public async Task NotifyMovieChange()
-    {
-        await Clients.All.SendAsync("ReceiveMovieUpdate");
-    }
-}
