+++ TestBlazor/Hubs/IMovieClient.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using TestBlazor.Models;
+
+namespace TestBlazor.Hubs
+{
+    public interface IMovieClient
+    {
+        Task MovieChanged(Movie movie);
+    }
+}
