+++ TestBlazor/Hubs/MovieHub.cs
@@ -1,0 +1,13 @@
+using Microsoft.AspNetCore.SignalR;
+using TestBlazor.Models;
+
+namespace TestBlazor.Hubs
+{
+    public class MovieHub : Hub<IMovieClient>
+    {
+        public async Task BroadcastMovieChange(Movie movie)
+        {
+            await Clients.All.MovieChanged(movie);
+        }
+    }
+}
